version 1.1

+++ myclub_website/events/views.py
@@ -4,7 +4,8 @@ from calendar import HTMLCalendar
 from datetime import datetime
 from django.http import HttpResponseRedirect
 from .models import Event, Venue
-from .forms import VenueForm, EventForm
+from .forms import VenueForm, EventForm, EventFormAdmin
+from django.contrib import messages
 
 #import pagination stuff
 from django.core.paginator import Paginator
@@ -20,13 +21,22 @@ def delete_venue(request, venue_id):
 #delete an event
 def delete_event(request, event_id):
 	event = Event.objects.get(pk=event_id)
-	event.delete()
-	return redirect('list-events')
+	if request.user == event.manager:
+		event.delete()
+		messages.success(request, ("Event Deleted"))
+		return redirect('list-events')
+	else:
+		messages.success(request, ("You are not authorized to Delete this event"))
+		return redirect('list-events')
 
 
 def update_event(request, event_id):
 	event = Event.objects.get(pk=event_id)
-	form = EventForm(request.POST or None, instance=event)
+	if request.user.is_superuser:
+		form = EventFormAdmin(request.POST or None, instance=event)
+	else:
+		form = EventForm(request.POST or None, instance=event)
+	
 	if form.is_valid():
 		form.save()
 		return redirect('list-events')
@@ -37,12 +47,26 @@ def update_event(request, event_id):
 def add_event(request):
 	submitted = False
 	if request.method == "POST":
-		form = EventForm(request.POST)
-		if form.is_valid():
-			form.save()
-			return HttpResponseRedirect('/add_event?submitted=True')
+		if request.user.is_superuser:
+			form = EventFormAdmin(request.POST)
+			if form.is_valid():
+				form.save()
+				return HttpResponseRedirect('/add_event?submitted=True')
+		else:
+			form = EventForm(request.POST)
+			if form.is_valid():
+				#form.save()
+				event = form.save(commit=False)
+				event.manager = request.user    #logged in user
+				event.save()
+				return HttpResponseRedirect('/add_event?submitted=True')
 	else:
-		form = EventForm
+		#just going to the page, not submitting
+		if request.user.is_superuser:
+			form = EventFormAdmin
+		else:
+			form = EventForm
+
 		if 'submitted' in request.GET:
 			submitted = True
 			
@@ -79,7 +103,7 @@ def list_venues(request):
 	venue_list = Venue.objects.all().order_by('name')
 
 	#set up pagination
-	p = Paginator(Venue.objects.all(), 1)
+	p = Paginator(Venue.objects.all(), 10)
 	page = request.GET.get('page')
 	venues = p.get_page(page)
 	nums = "a" * venues.paginator.num_pages
@@ -92,7 +116,10 @@ def add_venue(request):
 	if request.method == "POST":
 		form = VenueForm(request.POST)
 		if form.is_valid():
-			form.save()
+			venue = form.save(commit=False)
+			venue.owner = request.user.id    #logged in user
+			venue.save()
+			#form.save()
 			return HttpResponseRedirect('/add_venue?submitted=True')
 	else:
 		form = VenueForm
@@ -106,7 +133,7 @@ def all_events(request):
 	event_list = Event.objects.all().order_by('event_date')
 
 	#set up pagination
-	p = Paginator(Event.objects.all(), 2)
+	p = Paginator(Event.objects.all(), 10)
 	page = request.GET.get('page')
 	events = p.get_page(page)
 	nums = "a" * events.paginator.num_pages
